@@ -1,45 +1,50 @@
 {
-  "name": "wtfimnss-ui",
-  "version": "0.1.0",
+  "name": "html-quasar",
+  "version": "0.0.1",
+  "description": "WTF user interface",
+  "productName": "WTF-UI",
+  "author": "Lewis Pringle <lewis@sophists.com>",
   "private": true,
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint",
-    "debug": "vue-cli-service build --mode debug"
+    "dev": "quasar dev",
+    "build": "quasar build",
+    "lint": "eslint --ext .js,.ts,.vue ./",
+    "format": "prettier --write \"**/*.{js,ts,vue,scss,html,md,json}\" --ignore-path .gitignore",
+    "test": "echo \"No test specified\" && exit 0"
   },
   "dependencies": {
-    "core-js": "^3.20.1",
-    "vue": "^2.6.12",
-    "vue-class-component": "^6.0.0",
-    "vue-filter-pretty-bytes": "^0.1.6",
-    "vue-json-viewer": "^2.2.16",
-    "vue-moment": "^4.1.0",
-    "vue-property-decorator": "^7.0.0",
-    "vue-router": "^3.4.9",
-    "vuetify": "^2.4.0",
-    "vuex": "^3.6.0"
+    "@quasar/extras": "^1.0.0",
+    "@vueuse/core": "^8.7.4",
+    "json-viewer": "^2.2.0",
+    "moment": "^2.29.3",
+    "pinia": "^2.0.11",
+    "quasar": "^2.6.0",
+    "vue": "^3.0.0",
+    "vue-json-viewer": "^3.0.4",
+    "vue-router": "^4.0.0"
   },
   "devDependencies": {
-    "@types/vue-moment": "^4.0.2",
-    "@vue/cli-plugin-babel": "^4.5.9",
-    "@vue/cli-plugin-typescript": "^4.5.9",
-    "@vue/cli-service": "^4.5.9",
-    "sass": "^1.45.1",
-    "sass-loader": "^8.0.0",
+    "@quasar/app-vite": "^1.0.0",
+    "@types/node": "^12.20.21",
+    "@typescript-eslint/eslint-plugin": "^5.10.0",
+    "@typescript-eslint/parser": "^5.10.0",
+    "autoprefixer": "^10.4.2",
+    "eslint": "^8.10.0",
+    "eslint-config-prettier": "^8.1.0",
+    "eslint-plugin-vue": "^8.5.0",
+    "prettier": "^2.5.1",
     "typescript": "^4.5.4",
-    "vue-cli-plugin-vuetify": "^2.0.8",
-    "vue-template-compiler": "^2.6.12",
-    "vuetify-loader": "^1.6.0"
+    "workbox-build": "^6.5.3",
+    "workbox-cacheable-response": "^6.5.3",
+    "workbox-core": "^6.5.3",
+    "workbox-expiration": "^6.5.3",
+    "workbox-precaching": "^6.5.3",
+    "workbox-routing": "^6.5.3",
+    "workbox-strategies": "^6.5.3"
   },
-  "postcss": {
-    "plugins": {
-      "autoprefixer": {}
-    }
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "not ie <= 8"
-  ]
+  "engines": {
+    "node": "^18 || ^16 || ^14.19",
+    "npm": ">= 6.13.4",
+    "yarn": ">= 1.21.1"
+  }
 }
