@@ -1,305 +1,274 @@
-<template>
-  <div>
-    <table class="detailsTable" v-bind:key="device.id">
-      <tr>
-        <td class="labelColumn">Name</td>
-        <td>{{ device.name }}</td>
-      </tr>
-      <tr>
-        <td class="labelColumn">ID</td>
-        <td>
-          {{ device.id }}
-          <span class="snapshot" v-if="device.historicalSnapshot == true">{snapshot}</span>
-        </td>
-      </tr>
-      <tr v-if="device.type">
-        <td class="labelColumn">Types</td>
-        <td>{{ device.type.join(", ") }}</td>
-      </tr>
-      <tr v-if="device.icon">
-        <td>Icon</td>
-        <td><img :src="device.icon" width="24" height="24" /></td>
-      </tr>
-      <tr v-if="device.manufacturer">
-        <td class="labelColumn">Manufacturer</td>
-        <td>
-          <span v-if="device.manufacturer.shortName || device.manufacturer.fullName">{{
-            device.manufacturer.shortName || device.manufacturer.fullName
-          }}</span>
-          <span v-if="device.manufacturer.webSiteURL">
-            <span v-if="device.manufacturer.shortName || device.manufacturer.fullName">; </span>
-            Link:
-            <a :href="device.manufacturer.webSiteURL" target="_blank">{{
-              device.manufacturer.webSiteURL
-            }}</a>
-          </span>
-        </td>
-      </tr>
-      <tr v-if="device.operatingSystem">
-        <td class="labelColumn">OS</td>
-        <td>
-          {{ device.operatingSystem.fullVersionedName }}
-        </td>
-      </tr>
-      <tr v-if="device.lastSeenAt">
-        <td class="labelColumn">Last Seen</td>
-        <td>{{ device.lastSeenAt | moment("from", "now") }}</td>
-      </tr>
-      <tr>
-        <td class="labelColumn">Networks</td>
-        <td>
-          <table>
-            <tr
-              v-for="attachedNetID in Object.keys(device.attachedNetworks)"
-              v-bind:key="attachedNetID"
-            >
-              <td valign="top">&#x25cf;</td>
-              <td>
-                <table>
-                  <tr>
-                    <td>Name (ID)</td>
-                    <td class="nowrap">
-                      <ReadOnlyTextWithHover
-                        :message="
-                          deviceDetails.attachedNetworks[attachedNetID].name +
-                            ' (' +
-                            attachedNetID +
-                            ')'
-                        "
-                        :link="GetNetworkLink(attachedNetID)"
-                      />
-                    </td>
-                  </tr>
-                  <tr v-if="device.attachedNetworks[attachedNetID].hardwareAddresses">
-                    <td>Hardware Address(es)</td>
-                    <td class="nowrap">
-                      {{ device.attachedNetworks[attachedNetID].hardwareAddresses.join(", ") }}
-                    </td>
-                  </tr>
-                  <tr v-if="device.attachedNetworks[attachedNetID].localAddresses">
-                    <td>Network Address Binding(s)</td>
-                    <td class="nowrap">
-                      {{ device.attachedNetworks[attachedNetID].localAddresses.join(", ") }}
-                    </td>
-                  </tr>
-                </table>
-              </td>
-            </tr>
-          </table>
-        </td>
-      </tr>
-      <tr>
-        <td class="labelColumn">Services</td>
-        <td>
-          <table>
-            <tr v-for="svc in GetServices(device)">
-              <td>
-                <img
-                  v-if="ComputeServiceTypeIconURL(svc.name).url"
-                  :src="ComputeServiceTypeIconURL(svc.name).url"
-                  height="20"
-                  width="20"
-                />
-              </td>
-              <td class="labelColumn">{{ svc.name }}</td>
-              <td>
-                <a v-for="l in svc.links" v-bind:href="l.href" class="list-items" target="_blank">{{
-                  l.href
-                }}</a>
-              </td>
-            </tr>
-            <tr v-if="GetServices(device).length == 0">
-              <td><em>none</em></td>
-            </tr>
-          </table>
-        </td>
-      </tr>
-      <tr>
-        <td class="labelColumn">Open Ports</td>
-        <td>
-          <v-btn
-            class="smallBtnMargin"
-            elevation="2"
-            x-small
-            @click="rescanDevice"
-            :disabled="isRescanning"
-            >Rescan</v-btn
-          >
-          <span v-if="device.openPorts">{{ device.openPorts.join(", ") }}</span>
-        </td>
-      </tr>
-      <tr v-if="device.attachedNetworkInterfaces">
-        <td class="labelColumn">ATTACHED NETWORK INTERFACES</td>
-        <td>
-          <json-viewer :value="device.attachedNetworkInterfaces" :expand-depth="0" copyable sort />
-        </td>
-      </tr>
-      <tr v-if="device.aggregatesReversibly && device.aggregatesReversibly.length">
-        <td>Aggregates Reversibly</td>
-        <td>
-          <span v-for="aggregate in device.aggregatesReversibly" v-bind:key="aggregate">
-            <ReadOnlyTextWithHover :message="aggregate" :link="'/#/device/' + aggregate" />;
-          </span>
-        </td>
-      </tr>
-      <tr v-if="device.aggregatesIrreversibly && device.aggregatesIrreversibly.length">
-        <td>Aggregates Irreversibly</td>
-        <td>
-          <span v-for="aggregate in device.aggregatesIrreversibly" v-bind:key="aggregate">
-            <ReadOnlyTextWithHover :message="aggregate" />;
-          </span>
-        </td>
-      </tr>
-      <tr v-if="device.debugProps">
-        <td class="labelColumn">DEBUG INFO</td>
-        <td>
-          <json-viewer :value="device.debugProps" :expand-depth="1" copyable sort></json-viewer>
-        </td>
-      </tr>
-    </table>
-  </div>
-</template>
+<script setup lang="ts">
+import { defineProps, defineComponent, onMounted, onUnmounted, ref, computed } from 'vue';
+import * as moment from 'moment';
+import JsonViewer from 'vue-json-viewer';
 
-<script lang="ts">
-import { IDevice, INetworkAttachmentInfo } from "@/models/device/IDevice";
-import {
-  ComputeDeviceTypeIconURLs,
-  ComputeOSIconURLList,
-  ComputeServiceTypeIconURL,
-} from "@/models/device/Utils";
-import { INetwork } from "@/models/network/INetwork";
-import { GetNetworkLink, GetNetworkName, GetServices } from "@/models/network/Utils";
-import { OperatingSystem } from "@/models/OperatingSystem";
-import { rescanDevice } from "@/proxy/API";
-import { Component, Prop, Vue, Watch } from "vue-property-decorator";
+import { IDevice, INetworkAttachmentInfo } from "../models/device/IDevice";
+import { ComputeServiceTypeIconURL } from "../models/device/Utils";
+import { GetNetworkLink, GetNetworkName, GetServices } from "../models/network/Utils";
+import { rescanDevice } from "../proxy/API";
 
-@Component({
-  name: "DeviceDetails",
-  components: {
-    ReadOnlyTextWithHover: () => import("@/components/ReadOnlyTextWithHover.vue"),
-  },
+
+import ReadOnlyTextWithHover from '../components/ReadOnlyTextWithHover.vue';
+import Link2DetailsPage from '../components/Link2DetailsPage.vue';
+
+import { useNetStateStore } from '../stores/Net-State-store'
+
+const store = useNetStateStore()
+
+const props = defineProps({
+  deviceId: { type: String, required: true },
+  includeLinkToDetailsPage: { type: Boolean, required: false, default: false },
 })
-export default class DeviceDetails extends Vue {
-  @Prop({
-    required: true,
-    default: null,
-  })
-  public deviceId!: string;
 
-  private polling: undefined | number = undefined;
-  private isRescanning: boolean = false;
+defineComponent({
+  components: {
+    ReadOnlyTextWithHover,
+    Link2DetailsPage,
+    JsonViewer,
+  },
+});
 
-  private GetNetworkName = GetNetworkName;
-  private GetNetworkLink = GetNetworkLink;
-  private GetServices = GetServices;
-  private ComputeServiceTypeIconURL = ComputeServiceTypeIconURL;
+let polling: undefined | NodeJS.Timeout;
+var isRescanning: boolean = false;
 
-  private get localNetworkAddresses(): string[] {
-    const addresses: string[] = [];
-    if (this.device) {
-      Object.entries(this.device.attachedNetworks).forEach((element) => {
-        element[1].localAddresses.forEach((e: string) => addresses.push(e));
-      });
+function localNetworkAddresses(): string[] {
+  const addresses: string[] = [];
+  if (currentDevice.value) {
+    Object.entries(currentDevice.value.attachedNetworks).forEach((element) => {
+      element[1].localAddresses.forEach((e: string) => addresses.push(e));
+    });
+  }
+  return addresses.filter((value, index, self) => self.indexOf(value) === index);
+}
+
+onMounted(() => {
+  // first time check immediately, then more gradually for updates
+  store.fetchDevice(props.deviceId);
+  if (currentDevice.value) {
+    store.fetchNetworks(Object.keys(currentDevice.value.attachedNetworks));
+  } else {
+    store.fetchAvailableNetworks();
+  }
+  if (polling) {
+    clearInterval(polling);
+  }
+  polling = setInterval(() => {
+    store.fetchDevice(props.deviceId);
+    if (currentDevice.value) {
+      store.fetchNetworks(Object.keys(currentDevice.value.attachedNetworks));
     }
-    return addresses.filter((value, index, self) => self.indexOf(value) === index);
-  }
+  }, 15 * 1000);
+})
 
-  private created() {
-    this.pollData();
-  }
+onUnmounted(() => {
+  clearInterval(polling);
+})
 
-  private beforeDestroy() {
-    clearInterval(this.polling);
+async function rescanSelectedDevice(): Promise<void> {
+  isRescanning = true;
+  try {
+    await rescanDevice(props.deviceId);
+    store.fetchDevice(props.deviceId);
+  } finally {
+    isRescanning = false;
   }
+}
 
-  private get networks(): INetwork[] {
-    // @todo fix to not use getavailabletnetwks but fetch just the right ones
-    return this.$store.getters.getAvailableNetworks;
-  }
+const currentDevice = computed<IDevice | undefined>(
+  () => store.getDevice(props.deviceId)
+)
 
-  private async rescanDevice() {
-    this.isRescanning = true;
-    try {
-      await rescanDevice(this.deviceId);
-      this.$store.dispatch("fetchDevice", this.deviceId);
-    } finally {
-      this.isRescanning = false;
-    }
-  }
 
-  private pollData() {
-    // first time check immediately, then more gradually for updates
-    this.$store.dispatch("fetchDevice", this.deviceId);
-    if (this.device) {
-      this.$store.dispatch("fetchNetworks", Object.keys(this.device.attachedNetworks));
-    } else {
-      this.$store.dispatch("fetchAvailableNetworks");
-    }
-    if (this.polling) {
-      clearInterval(this.polling);
-    }
-    this.polling = setInterval(() => {
-      this.$store.dispatch("fetchDevice", this.deviceId);
-      if (this.device) {
-        this.$store.dispatch("fetchNetworks", Object.keys(this.device.attachedNetworks));
-      }
-    }, 15 * 1000);
-  }
+interface IExtendedDevice extends IDevice {
+  localAddresses: string;
+  attachedNetworks: any;
+}
 
-  private get device(): IDevice | undefined {
-    return this.$store.getters.getDevice(this.deviceId);
-  }
-
-  private get deviceDetails(): object | undefined {
-    const d = this.device;
-    if (d) {
+let currentDeviceDetails = computed<IExtendedDevice | undefined>(
+  () => {
+    if (currentDevice.value) {
       const attachedNetworkInfo = {} as { [key: string]: object };
-      Object.keys(d.attachedNetworks).forEach((element: any) => {
-        const thisNWI = d.attachedNetworks[element] as INetworkAttachmentInfo;
+      Object.keys(currentDevice.value.attachedNetworks).forEach((element: any) => {
+        const thisNWI = currentDevice.value.attachedNetworks[element] as INetworkAttachmentInfo;
         let netName = "?";
-        const thisNetObj = this.$store.getters.getNetwork(element);
+        const thisNetObj = store.getNetwork(element);
         if (thisNetObj) {
           netName = GetNetworkName(thisNetObj);
         }
         attachedNetworkInfo[element] = { ...thisNWI, name: netName };
       });
       return {
-        ...d,
+        ...currentDevice.value,
         ...{
-          localAddresses: this.localNetworkAddresses.join(", "),
+          localAddresses: localNetworkAddresses().join(", "),
           attachedNetworks: attachedNetworkInfo,
         },
       };
     }
     return undefined;
   }
-}
+)
 </script>
+
+<template>
+  <div v-if="currentDevice" class="q-pa-sm">
+    <Link2DetailsPage :link="'/#/device/' + currentDevice.id" v-if="props.includeLinkToDetailsPage" />
+
+    <div class="row">
+      <div class="col-3 labelColumn"> Name </div>
+      <div class="col"> {{ currentDevice.name }} </div>
+    </div>
+    <div class="row">
+      <div class="col-3 labelColumn"> ID </div>
+      <div class="col"> {{ currentDevice.id }} <span class="snapshot"
+          v-if="currentDevice.historicalSnapshot == true">{snapshot}</span></div>
+    </div>
+    <div class="row" v-if="currentDevice.type">
+      <div class="col-3 labelColumn"> Types </div>
+      <div class="col"> {{ currentDevice.type.join(", ") }}</div>
+    </div>
+    <div class="row" v-if="currentDevice.icon">
+      <div class="col-3 labelColumn"> Icon </div>
+      <div class="col"> <img :src="currentDevice.icon" width="24" height="24" /></div>
+    </div>
+    <div class="row" v-if="currentDevice.manufacturer">
+      <div class="col-3 labelColumn"> Manufacturer </div>
+      <div class="col">
+        <span v-if="currentDevice.manufacturer.shortName || currentDevice.manufacturer.fullName">{{
+            currentDevice.manufacturer.shortName || currentDevice.manufacturer.fullName
+        }}</span>
+        <span v-if="currentDevice.manufacturer.webSiteURL">
+          <span v-if="currentDevice.manufacturer.shortName || currentDevice.manufacturer.fullName">; </span>
+          Link:
+          <a :href="currentDevice.manufacturer.webSiteURL" target="_blank">{{
+              currentDevice.manufacturer.webSiteURL
+          }}</a>
+        </span>
+      </div>
+    </div>
+    <div class="row" v-if="currentDevice.operatingSystem">
+      <div class="col-3 labelColumn"> OS </div>
+      <div class="col"> {{ currentDevice.operatingSystem.fullVersionedName }}</div>
+    </div>
+    <div class="row" v-if="currentDevice.lastSeenAt">
+      <div class="col-3 labelColumn"> Last Seen </div>
+      <div class="col"> {{ moment(currentDevice.lastSeenAt).fromNow() }}</div>
+    </div>
+    <div class="row">
+      <div class="col-3 labelColumn"> Networks </div>
+      <div class="col">
+        <div class="row" v-for="attachedNetID in Object.keys(currentDevice.attachedNetworks)"
+          v-bind:key="attachedNetID">
+          <div class="col">
+            <div class="row">
+              <div class="col no-wrap">
+                <ReadOnlyTextWithHover :message="
+                  currentDeviceDetails.attachedNetworks[attachedNetID].name +
+                  ' (' +
+                  attachedNetID +
+                  ')'
+                " :link="GetNetworkLink(attachedNetID)" title="Network Name" />
+              </div>
+            </div>
+            <div class="row" v-if="currentDevice.attachedNetworks[attachedNetID].hardwareAddresses">
+              <div class="col-1" />
+              <div class="col-4">Hardware Address(es)</div>
+              <div class="col no-wrap"> {{ currentDevice.attachedNetworks[attachedNetID].hardwareAddresses.join(", ") }}
+              </div>
+            </div>
+            <div class="row" v-if="currentDevice.attachedNetworks[attachedNetID].localAddresses">
+              <div class="col-1" />
+              <div class="col-4">Network Address Binding(s)</div>
+              <div class="col no-wrap"> {{ currentDevice.attachedNetworks[attachedNetID].localAddresses.join(", ") }}
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+    <div class="row">
+      <div class="col-3 labelColumn"> Services </div>
+      <div class="col">
+        <div class="row" v-for="svc in GetServices(currentDevice)" v-bind:key="svc.name">
+          <div class="col-1">
+            <img v-if="ComputeServiceTypeIconURL(svc.name).url" :src="ComputeServiceTypeIconURL(svc.name).url"
+              height="20" width="20" />
+          </div>
+          <div class="col-1">{{ svc.name }}</div>
+          <div class="col no-wrap"> <a v-for="l in svc.links" v-bind:href="l.href" v-bind:key="l.href"
+              class="list-items" target="_blank">{{
+                  l.href
+              }}</a>
+          </div>
+        </div>
+      </div>
+    </div>
+    <div class="row">
+      <div class="col-3 labelColumn"> Open Ports </div>
+      <div class="col">
+        <q-btn class="smallBtnMargin" elevation="2" dense x-small @click="rescanSelectedDevice"
+          :disabled="isRescanning"> Rescan </q-btn>
+        <span v-if="currentDevice.openPorts">{{ currentDevice.openPorts.join(", ") }}</span>
+      </div>
+    </div>
+    <div class="row" v-if="currentDevice.attachedNetworkInterfaces">
+      <div class="col-3 labelColumn"> ATTACHED NETWORK INTERFACES </div>
+      <div class="col">
+        <json-viewer :value="currentDevice.attachedNetworkInterfaces" :expand-depth="0" copyable sort
+          class="debugInfoJSONViewers" />
+      </div>
+    </div>
+    <div class="row" v-if="currentDevice.aggregatesReversibly && currentDevice.aggregatesReversibly.length">
+      <div class="col-3 labelColumn"> Aggregates Reversibly </div>
+      <div class="col">
+        <span v-for="aggregate in currentDevice.aggregatesReversibly" v-bind:key="aggregate">
+          <ReadOnlyTextWithHover :message="aggregate" :link="'/#/device/' + aggregate" />;
+        </span>
+      </div>
+    </div>
+    <div class="row" v-if="currentDevice.aggregatesIrreversibly && currentDevice.aggregatesIrreversibly.length">
+      <div class="col-3 labelColumn"> Aggregates Irreversibly </div>
+      <div class="col">
+        <span v-for="aggregate in currentDevice.aggregatesIrreversibly" v-bind:key="aggregate">
+          <ReadOnlyTextWithHover :message="aggregate" />;
+        </span>
+      </div>
+    </div>
+    <div class="row" v-if="currentDevice.debugProps">
+      <div class="col-3 labelColumn"> DEBUG INFO </div>
+      <div class="col">
+        <json-viewer :value="currentDevice.debugProps" :expand-depth="0" copyable sort class="debugInfoJSONViewers" />
+      </div>
+    </div>
+  </div>
+</template>
 
 <style scoped lang="scss">
 .list-items {
   padding-right: 1em;
 }
-td.labelColumn {
+
+.labelColumn {
   vertical-align: top;
 }
-.detailsTable {
-  table-layout: fixed;
-}
-.detailsTable td {
-  padding-left: 5px;
-  padding-right: 10px;
-}
+
 .nowrap {
   white-space: nowrap;
   overflow: hidden;
   text-overflow: ellipsis;
 }
+
 .smallBtnMargin {
   margin-left: 1em;
   margin-right: 1em;
 }
+
 .snapshot {
   font-style: italic;
+}
+
+.debugInfoJSONViewers {
+  margin-right: 2em;
 }
 </style>
