@@ -1,52 +1,40 @@
-<template>
-  <span>{{ msg }}</span>
-</template>
+<script setup lang="ts">
+import { onMounted, defineProps, watch, ref } from 'vue';
 
-<script lang="ts">
-import { Component, Prop, Vue, Watch } from "vue-property-decorator";
+const props = defineProps({
+  nTotalItems: { type: Number, required: true },
+  filtered: { type: Boolean, default: false },
+  nItemsSelected: { type: Number, required: true, },
+  itemsName: { type: String, default: "items" },
+})
+
+function onChange() {
+  let filtered = props.filtered;
+  if (filtered === null) {
+    filtered = props.nItemsSelected === props.nTotalItems;
+  }
+  if (filtered) {
+    msg.value = `Filtered: showing ${props.nItemsSelected} of ${props.nTotalItems} ${props.itemsName}`;
+  } else {
+    msg.value = `Unfiltered: all ${props.nTotalItems} ${props.itemsName} showing`;
+  }
+}
+
+onMounted(() => {
+  onChange()
+})
+
+watch([() => props.filtered, () => props.nTotalItems, () => props.nItemsSelected, () => props.itemsName], onChange)
 
 /*
  *  This is for use in the filter section of the app-bar, to say how much is filtered out.
  */
-@Component({
-  name: "FilterSummaryMessage",
-})
-export default class FilterSummaryMessage extends Vue {
-  @Prop({ required: true })
-  public nTotalItems!: number;
-
-  @Prop({ required: false })
-  public filtered!: boolean;
-
-  @Prop({ required: true })
-  public nItemsSelected!: number;
-
-  @Prop({ default: "items" })
-  public itemsName!: string;
-
-  private msg: string = "";
-
-  private created() {
-    this.onChange();
-  }
-
-  @Watch("filtered")
-  @Watch("nTotalItems")
-  @Watch("nItemsSelected")
-  @Watch("itemsName")
-  private onChange() {
-    let filtered = this.filtered;
-    if (filtered === null) {
-      filtered = this.nItemsSelected === this.nTotalItems;
-    }
-    if (filtered) {
-      this.msg = `Filtered: showing ${this.nItemsSelected} of ${this.nTotalItems} ${this.itemsName}`;
-    } else {
-      this.msg = `Unfiltered: all ${this.nTotalItems} ${this.itemsName} showing`;
-    }
-  }
-}
+var msg = ref("");
 </script>
+
+<template>
+  <span>{{ msg }}</span>
+</template>
 
 <style scoped lang="scss">
 div {
