@@ -1,5 +1,12 @@
-declare module "*.vue" {
-  import Vue from "vue";
-  export default Vue;
+/* eslint-disable */
+
+/// <reference types="vite/client" />
+/// <reference types="vue/ref-macros" />
+
+// Mocks all files ending in `.vue` showing them as plain Vue instances
+declare module '*.vue' {
+  import type { DefineComponent } from 'vue'
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any, @typescript-eslint/ban-types
+  const component: DefineComponent<{}, {}, any>
+  export default component
 }
-declare module "vue-filter-pretty-bytes";
